feat: rebuild /auth/verify-request on the shared auth-surface design
The verify-request page was still the old Mantine split layout, wrapped
in the app Header by the (sidemenu) layout. Move it to the standalone
/auth root layout and rebuild it on the same .auth-surface chrome as
/signin, with the left column carrying the check-your-email content
(email + sign-in code fields, expiry and spam notes, back link).

The /signin marketing aside is extracted into a shared
AuthMarketingPanel component consumed by both pages so they cannot
drift. Form behavior is unchanged: sessionStorage recovery, focus
handoff, code normalization, and the always-visible email field.

The URL stays /auth/verify-request, so middleware, auth config, and
the pages that navigate here are untouched.
